// frontend/src/lib/api/generated/models/productResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Nethub POS MVP
 * OpenAPI spec version: v0.0.1
 */
import type { BaseAttributes } from "./baseAttributes";

export interface ProductResponse {
  id: string;
  label: string;
  selling_price: number;
  track_stock: boolean;
  stock: number;
  active: boolean;
  attributes: BaseAttributes;
}